Customer plan backend

+++ src/main/java/com/Estructura/API/repository/CustomerRequestRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface CustomerRequestRepository extends JpaRepository<CustomerRequest,Integer> {
+public interface CustomerRequestRepository extends JpaRepository<CustomerRequest,Long> {
     List<CustomerRequest> findByTargetCategoriesRole(Role role);
     Optional<CustomerRequest> findCustomerRequestById(Long id);
